hci_usb: Avoid writing more data to USB than max packet size supported
This is a hack to get btusb working.
This should be rather fixed in the USB device stack.

try_write() writes the whole net buffer to USB which sometimes
is larger than the maximum packet size of an endpoint.

To workaround this, use the ret_bytes param to know how much
has been written to the USB bus and write the remaining chunks
in the subsequent calls.

Signed-off-by: Sundar Subramaniyan <sundar.subramaniyan@gmail.com>