Bluetooth: Allow tx_pwr to be read when not BT_CONN
The beacon sample, a connection-less BT sample, is failing. It fails
because hci.c assumes there exists a function for reading out the TX
power, but ll_sw/ctrl.c does not define one when not BT_CONN.

We fix this by moving ll_tx_pwr_lvl_get such that it is defined when
not BT_CONN. This is safe because all ll_tx_pwr_lvl_get does is write
a static value to an output variable. It was also checking for a valid
connection, but this is not necessary, so this has been dropped.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>